Player and Bot in new file!
