+++ standard_names_SH.py
@@ -22,13 +22,6 @@ class X:
 
 # class HTML(str):
 #     pass
-
-class Player:
-    pass
-class Bot:
-    pass
-class Log:
-    pass
 class LogInfo:
     pass
 class MyErr(BaseException):
